// apps/backend/src/modules/authorization/enums/permission.ts
export enum Permission {
    TOWNHOUSE_CREATE = 'TOWNHOUSE_CREATE',
    TOWNHOUSE_READ = 'TOWNHOUSE_READ',
    TOWNHOUSE_UPDATE = 'TOWNHOUSE_UPDATE',
    TOWNHOUSE_DELETE = 'TOWNHOUSE_DELETE',
    HOUSE_CREATE = 'HOUSE_CREATE',
    HOUSE_READ = 'HOUSE_READ',
    HOUSE_UPDATE = 'HOUSE_UPDATE',
    HOUSE_DELETE = 'HOUSE_DELETE',
    USER_CREATE = 'USER_CREATE',
    USER_READ = 'USER_READ',
    USER_UPDATE = 'USER_UPDATE',
    USER_PROVISIONAL_PASSWORD_CREATE = 'USER_PROVISIONAL_PASSWORD_CREATE',
    USER_ROLE_UPDATE = 'USER_ROLE_UPDATE',
    CONTRIBUTION_CREATE = 'CONTRIBUTION_CREATE',
    CONTRIBUTION_READ = 'CONTRIBUTION_READ',
    CONTRIBUTION_UPDATE = 'CONTRIBUTION_UPDATE',
    CONTRIBUTION_FINALIZE = 'CONTRIBUTION_FINALIZE',
    CONTRIBUTION_REOPEN = 'CONTRIBUTION_REOPEN',
    CONTRIBUTION_CANCEL = 'CONTRIBUTION_CANCEL',
    PAYMENT_PROOF_CREATE = 'PAYMENT_PROOF_CREATE',
    PAYMENT_PROOF_READ = 'PAYMENT_PROOF_READ',
    ACCOUNTABILITY_CREATE = 'ACCOUNTABILITY_CREATE',
    ACCOUNTABILITY_READ = 'ACCOUNTABILITY_READ',
    AUDIT_READ = 'AUDIT_READ',
}